Remove param `stylesPrepend` from `lazySet`
