+++ lib/flutter_widget_from_html.dart
@@ -1,3 +1,4 @@
+export 'src/builder.dart';
 export 'src/data_classes.dart';
 export 'src/helpers.dart';
 export 'src/html_widget.dart';
