+++ lib/src/builder.dart
@@ -1,0 +1,1 @@
+export 'package:flutter_widget_from_html_core/src/builder.dart';
