+++ packages/core/lib/flutter_widget_from_html_core.dart
@@ -1,3 +1,4 @@
+export 'src/builder.dart';
 export 'src/core_helpers.dart';
 export 'src/core_html_widget.dart';
 export 'src/core_widget_factory.dart';
