+++ packages/core/lib/src/ops/style_bg_color.dart
@@ -17,13 +17,13 @@ class _StyleBgColor {
   BuildOp get buildOp => BuildOp(
       isBlockElement: false,
       onPieces: (meta, pieces) {
-        final bgColor = _parseColor(meta);
+        final bgColor = wf.parseColor(meta.style(_kCssBackgroundColor));
         if (bgColor == null) return pieces;
 
         return pieces.map((p) => p.hasWidgets ? p : _buildBlock(p, bgColor));
       },
       onWidgets: (meta, widgets) {
-        final bgColor = _parseColor(meta);
+        final bgColor = wf.parseColor(meta.style(_kCssBackgroundColor));
         if (bgColor == null) return null;
 
         return _listOrNull(_buildBox(widgets, bgColor));
@@ -35,12 +35,4 @@ class _StyleBgColor {
 
   Widget _buildBox(Iterable<Widget> widgets, Color bgColor) =>
       wf.buildDecoratedBox(wf.buildBody(widgets), color: bgColor);
-
-  Color _parseColor(NodeMetadata meta) {
-    String value;
-    meta.styles((k, v) => k == _kCssBackgroundColor ? value = v : null);
-    if (value == null) return null;
-
-    return wf.parseColor(value);
-  }
 }
